requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
attrs==23.2.0
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
Flask==3.0.2
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
multidict==6.0.5
openai==0.28.0
pydantic==2.6.1
pydantic_core==2.16.2
python-dotenv==1.0.1
requests==2.31.0
sniffio==1.3.0
tqdm==4.66.1
typing_extensions==4.9.0
urllib3==2.2.0
Werkzeug==3.0.1
yarl==1.9.4
